+++ src/pages/_app.tsx
@@ -3,7 +3,7 @@ import "@fontsource/hanken-grotesk";
 import type { AppProps } from "next/app";
 import { ChakraProvider } from "@chakra-ui/react";
 import { Layout } from "@/components/Layout";
-import theme from "./theme";
+import theme from "../styles/theme";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
